fixed #103 the Get labourerProfile and ClientProfile bugs

+++ backEnd/roleDemo/Repositories/ClientProfileVMRepo.cs
@@ -27,7 +27,7 @@ namespace labourRecruitment.Repositories
             ClientProfileVM cp = new ClientProfileVM()
             {
                 Client = Client,
-                AverageRating = (double)avgerageQuality
+                AverageRating = avgerageQuality
                 
             };
             return cp;
